Extract conversation turn checkpoint module
