+++ crates/app/src/conversation/turn_coordinator.rs
@@ -12,9 +12,8 @@ use async_trait::async_trait;
 #[cfg(feature = "memory-sqlite")]
 use futures_util::FutureExt;
 use loongclaw_contracts::{AuditEventKind, ExecutionPlane, PlaneTier};
-use serde::{Deserialize, Serialize};
+use serde::Serialize;
 use serde_json::{Value, json};
-use sha2::{Digest, Sha256};
 #[cfg(feature = "memory-sqlite")]
 use tokio::runtime::Handle;
 use tokio::sync::Mutex;
@@ -35,9 +34,8 @@ use super::super::config::LoongClawConfig;
 use super::ConversationSessionAddress;
 use super::ProviderErrorMode;
 use super::analytics::{
-    SafeLaneEventSummary, TurnCheckpointProgressStatus as AnalyticsTurnCheckpointProgressStatus,
-    TurnCheckpointRecoveryAction, TurnCheckpointRepairManualReason, TurnCheckpointRepairPlan,
-    TurnCheckpointSessionState, build_turn_checkpoint_repair_plan, summarize_safe_lane_history,
+    SafeLaneEventSummary, TurnCheckpointRecoveryAction, build_turn_checkpoint_repair_plan,
+    summarize_safe_lane_history,
 };
 use super::context_engine::{AssembledConversationContext, ConversationContextEngine};
 use super::ingress::ConversationIngressContext;
@@ -79,6 +77,21 @@ use super::turn_budget::{
     EscalatingAttemptBudget, SafeLaneBackpressureBudget, SafeLaneContinuationBudgetDecision,
     SafeLaneFailureRouteReason, SafeLaneReplanBudget,
 };
+#[cfg(test)]
+use super::turn_checkpoint::TurnCheckpointResultKind;
+use super::turn_checkpoint::{
+    ContextCompactionOutcome, TurnCheckpointDiagnostics, TurnCheckpointFailure,
+    TurnCheckpointFailureStep, TurnCheckpointFinalizationProgress, TurnCheckpointIdentity,
+    TurnCheckpointProgressStatus, TurnCheckpointRecoveryAssessment,
+    TurnCheckpointRepairResumeInput, TurnCheckpointRequest, TurnCheckpointSnapshot,
+    TurnCheckpointStage, TurnCheckpointTailRepairOutcome, TurnCheckpointTailRepairReason,
+    TurnCheckpointTailRepairRuntimeProbe, TurnCheckpointTailRepairSource,
+    TurnCheckpointTailRepairStatus, TurnCheckpointTailRuntimeEligibility,
+    TurnFinalizationCheckpoint, TurnLaneExecutionSnapshot, TurnPreparationSnapshot,
+    TurnReplyCheckpoint, checkpoint_context_fingerprint_sha256, persist_turn_checkpoint_event,
+    persist_turn_checkpoint_event_value, restore_analytics_turn_checkpoint_progress_status,
+    turn_checkpoint_result_kind,
+};
 use super::turn_engine::{
     AppToolDispatcher, DefaultAppToolDispatcher, ProviderTurn, ToolBatchExecutionIntentStatus,
     ToolBatchExecutionTrace, ToolIntent, TurnEngine, TurnFailure, TurnFailureKind, TurnResult,
@@ -89,14 +102,16 @@ use super::turn_observer::{
     ConversationTurnToolEvent, build_observer_streaming_token_callback,
 };
 use super::turn_shared::{
-    ProviderTurnRequestAction, ReplyPersistenceMode, ReplyResolutionMode, ToolDrivenFollowupKind,
-    ToolDrivenFollowupPayload, ToolDrivenReplyBaseDecision, ToolDrivenReplyPhase,
-    build_tool_driven_followup_tail, build_tool_loop_guard_tail,
-    decide_provider_turn_request_action, format_approval_required_reply, next_conversation_turn_id,
-    reduce_followup_payload_for_model, request_completion_with_raw_fallback,
-    tool_driven_followup_payload, tool_loop_circuit_breaker_reply,
-    tool_result_contains_truncation_signal, user_requested_raw_tool_output,
+    ProviderTurnRequestAction, ReplyPersistenceMode, ToolDrivenFollowupPayload,
+    ToolDrivenReplyBaseDecision, ToolDrivenReplyPhase, build_tool_driven_followup_tail,
+    build_tool_loop_guard_tail, decide_provider_turn_request_action,
+    format_approval_required_reply, next_conversation_turn_id, reduce_followup_payload_for_model,
+    request_completion_with_raw_fallback, tool_driven_followup_payload,
+    tool_loop_circuit_breaker_reply, tool_result_contains_truncation_signal,
+    user_requested_raw_tool_output,
 };
+#[cfg(test)]
+use super::turn_shared::{ReplyResolutionMode, ToolDrivenFollowupKind};
 #[cfg(feature = "memory-sqlite")]
 use crate::session::recovery::{
     RECOVERY_EVENT_KIND, build_async_spawn_failure_recovery_payload,
@@ -112,308 +127,6 @@ use crate::session::repository::{
 
 #[derive(Default)]
 pub struct ConversationTurnCoordinator;
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum TurnCheckpointTailRepairStatus {
-    NoCheckpoint,
-    NotNeeded,
-    Repaired,
-    ManualRequired,
-}
-
-impl TurnCheckpointTailRepairStatus {
-    pub fn as_str(self) -> &'static str {
-        match self {
-            Self::NoCheckpoint => "no_checkpoint",
-            Self::NotNeeded => "not_needed",
-            Self::Repaired => "repaired",
-            Self::ManualRequired => "manual_required",
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum TurnCheckpointTailRepairReason {
-    NoCheckpoint,
-    NotNeeded,
-    Repaired,
-    CheckpointIdentityMissing,
-    SafeLaneBackpressureTerminalRequiresManualInspection,
-    SafeLaneSessionGovernorTerminalRequiresManualInspection,
-    CheckpointPreparationMalformed,
-    CheckpointPreparationMismatch,
-    CheckpointPreparationFingerprintMismatch,
-    CheckpointStateRequiresManualInspection,
-    VisibleTurnPairMissing,
-    CheckpointIdentityMismatch,
-}
-
-impl TurnCheckpointTailRepairReason {
-    pub fn as_str(self) -> &'static str {
-        match self {
-            Self::NoCheckpoint => "no_checkpoint",
-            Self::NotNeeded => "not_needed",
-            Self::Repaired => "repaired",
-            Self::CheckpointIdentityMissing => "checkpoint_identity_missing",
-            Self::SafeLaneBackpressureTerminalRequiresManualInspection => {
-                "safe_lane_backpressure_terminal_requires_manual_inspection"
-            }
-            Self::SafeLaneSessionGovernorTerminalRequiresManualInspection => {
-                "safe_lane_session_governor_terminal_requires_manual_inspection"
-            }
-            Self::CheckpointPreparationMalformed => "checkpoint_preparation_malformed",
-            Self::CheckpointPreparationMismatch => "checkpoint_preparation_mismatch",
-            Self::CheckpointPreparationFingerprintMismatch => {
-                "checkpoint_preparation_fingerprint_mismatch"
-            }
-            Self::CheckpointStateRequiresManualInspection => {
-                "checkpoint_state_requires_manual_inspection"
-            }
-            Self::VisibleTurnPairMissing => "visible_turn_pair_missing",
-            Self::CheckpointIdentityMismatch => "checkpoint_identity_mismatch",
-        }
-    }
-}
-
-impl From<TurnCheckpointRepairManualReason> for TurnCheckpointTailRepairReason {
-    fn from(reason: TurnCheckpointRepairManualReason) -> Self {
-        match reason {
-            TurnCheckpointRepairManualReason::CheckpointIdentityMissing => {
-                Self::CheckpointIdentityMissing
-            }
-            TurnCheckpointRepairManualReason::SafeLaneBackpressureTerminalRequiresManualInspection => {
-                Self::SafeLaneBackpressureTerminalRequiresManualInspection
-            }
-            TurnCheckpointRepairManualReason::SafeLaneSessionGovernorTerminalRequiresManualInspection => {
-                Self::SafeLaneSessionGovernorTerminalRequiresManualInspection
-            }
-            TurnCheckpointRepairManualReason::CheckpointStateRequiresManualInspection => {
-                Self::CheckpointStateRequiresManualInspection
-            }
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct TurnCheckpointTailRepairOutcome {
-    status: TurnCheckpointTailRepairStatus,
-    action: TurnCheckpointRecoveryAction,
-    source: Option<TurnCheckpointTailRepairSource>,
-    reason: TurnCheckpointTailRepairReason,
-    session_state: TurnCheckpointSessionState,
-    checkpoint_events: u32,
-    after_turn_status: Option<&'static str>,
-    compaction_status: Option<&'static str>,
-}
-
-impl TurnCheckpointTailRepairOutcome {
-    fn no_checkpoint() -> Self {
-        Self {
-            status: TurnCheckpointTailRepairStatus::NoCheckpoint,
-            action: TurnCheckpointRecoveryAction::None,
-            source: None,
-            reason: TurnCheckpointTailRepairReason::NoCheckpoint,
-            session_state: TurnCheckpointSessionState::NotDurable,
-            checkpoint_events: 0,
-            after_turn_status: None,
-            compaction_status: None,
-        }
-    }
-
-    pub(crate) fn from_summary(
-        status: TurnCheckpointTailRepairStatus,
-        action: TurnCheckpointRecoveryAction,
-        source: Option<TurnCheckpointTailRepairSource>,
-        reason: TurnCheckpointTailRepairReason,
-        summary: &super::analytics::TurnCheckpointEventSummary,
-    ) -> Self {
-        Self {
-            status,
-            action,
-            source,
-            reason,
-            session_state: summary.session_state,
-            checkpoint_events: summary.checkpoint_events,
-            after_turn_status: summary
-                .latest_after_turn
-                .map(format_analytics_turn_checkpoint_progress_status),
-            compaction_status: summary
-                .latest_compaction
-                .map(format_analytics_turn_checkpoint_progress_status),
-        }
-    }
-
-    fn repaired(
-        action: TurnCheckpointRecoveryAction,
-        summary: &super::analytics::TurnCheckpointEventSummary,
-        after_turn_status: TurnCheckpointProgressStatus,
-        compaction_status: TurnCheckpointProgressStatus,
-    ) -> Self {
-        Self {
-            status: TurnCheckpointTailRepairStatus::Repaired,
-            action,
-            source: Some(TurnCheckpointTailRepairSource::Runtime),
-            reason: TurnCheckpointTailRepairReason::Repaired,
-            session_state: summary.session_state,
-            checkpoint_events: summary.checkpoint_events,
-            after_turn_status: Some(format_turn_checkpoint_progress_status(after_turn_status)),
-            compaction_status: Some(format_turn_checkpoint_progress_status(compaction_status)),
-        }
-    }
-
-    pub fn status(&self) -> TurnCheckpointTailRepairStatus {
-        self.status
-    }
-
-    pub fn action(&self) -> TurnCheckpointRecoveryAction {
-        self.action
-    }
-
-    pub fn source(&self) -> Option<TurnCheckpointTailRepairSource> {
-        self.source
-    }
-
-    pub fn reason(&self) -> TurnCheckpointTailRepairReason {
-        self.reason
-    }
-
-    pub fn session_state(&self) -> TurnCheckpointSessionState {
-        self.session_state
-    }
-
-    pub fn checkpoint_events(&self) -> u32 {
-        self.checkpoint_events
-    }
-
-    pub fn after_turn_status(&self) -> Option<&'static str> {
-        self.after_turn_status
-    }
-
-    pub fn compaction_status(&self) -> Option<&'static str> {
-        self.compaction_status
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum TurnCheckpointTailRepairSource {
-    Summary,
-    Runtime,
-}
-
-impl TurnCheckpointTailRepairSource {
-    pub fn as_str(self) -> &'static str {
-        match self {
-            Self::Summary => "summary",
-            Self::Runtime => "runtime",
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct TurnCheckpointTailRepairRuntimeProbe {
-    action: TurnCheckpointRecoveryAction,
-    source: TurnCheckpointTailRepairSource,
-    reason: TurnCheckpointTailRepairReason,
-}
-
-impl TurnCheckpointTailRepairRuntimeProbe {
-    pub(crate) fn new(
-        action: TurnCheckpointRecoveryAction,
-        source: TurnCheckpointTailRepairSource,
-        reason: TurnCheckpointTailRepairReason,
-    ) -> Self {
-        Self {
-            action,
-            source,
-            reason,
-        }
-    }
-
-    pub fn action(&self) -> TurnCheckpointRecoveryAction {
-        self.action
-    }
-
-    pub fn source(&self) -> TurnCheckpointTailRepairSource {
-        self.source
-    }
-
-    pub fn reason(&self) -> TurnCheckpointTailRepairReason {
-        self.reason
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub(crate) struct TurnCheckpointRecoveryAssessment {
-    action: TurnCheckpointRecoveryAction,
-    source: TurnCheckpointTailRepairSource,
-    reason: Option<TurnCheckpointTailRepairReason>,
-}
-
-impl TurnCheckpointRecoveryAssessment {
-    pub(crate) fn from_summary(summary: &super::analytics::TurnCheckpointEventSummary) -> Self {
-        let repair_plan = build_turn_checkpoint_repair_plan(summary);
-        let reason = matches!(
-            repair_plan.action(),
-            TurnCheckpointRecoveryAction::InspectManually
-        )
-        .then(|| {
-            repair_plan
-                .manual_reason()
-                .map(TurnCheckpointTailRepairReason::from)
-                .unwrap_or(TurnCheckpointTailRepairReason::CheckpointStateRequiresManualInspection)
-        });
-        Self {
-            action: repair_plan.action(),
-            source: TurnCheckpointTailRepairSource::Summary,
-            reason,
-        }
-    }
-
-    pub fn action(self) -> TurnCheckpointRecoveryAction {
-        self.action
-    }
-
-    pub fn source(self) -> TurnCheckpointTailRepairSource {
-        self.source
-    }
-
-    pub fn reason(self) -> Option<TurnCheckpointTailRepairReason> {
-        self.reason
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) struct TurnCheckpointDiagnostics {
-    summary: super::analytics::TurnCheckpointEventSummary,
-    recovery: TurnCheckpointRecoveryAssessment,
-    runtime_probe: Option<TurnCheckpointTailRepairRuntimeProbe>,
-}
-
-impl TurnCheckpointDiagnostics {
-    pub(crate) fn new(
-        summary: super::analytics::TurnCheckpointEventSummary,
-        recovery: TurnCheckpointRecoveryAssessment,
-        runtime_probe: Option<TurnCheckpointTailRepairRuntimeProbe>,
-    ) -> Self {
-        Self {
-            summary,
-            recovery,
-            runtime_probe,
-        }
-    }
-
-    pub fn summary(&self) -> &super::analytics::TurnCheckpointEventSummary {
-        &self.summary
-    }
-
-    pub fn recovery(&self) -> TurnCheckpointRecoveryAssessment {
-        self.recovery
-    }
-
-    pub fn runtime_probe(&self) -> Option<&TurnCheckpointTailRepairRuntimeProbe> {
-        self.runtime_probe.as_ref()
-    }
-}
 
 #[derive(Debug, Clone, Copy, Default, PartialEq, Eq)]
 struct SafeLaneExecutionMetrics {
@@ -1052,7 +765,7 @@ impl ProviderTurnContinuePhase {
             Some(reply),
             self.request.clone(),
             Some(self.lane_execution.checkpoint()),
-            Some(turn_reply_checkpoint(&self.reply_phase)),
+            Some(TurnReplyCheckpoint::from_phase(&self.reply_phase)),
             TurnFinalizationCheckpoint::persist_reply(ReplyPersistenceMode::Success),
         )
     }
@@ -1107,206 +820,6 @@ impl ProviderTurnContinuePhase {
         )
         .await
     }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize)]
-struct TurnCheckpointSnapshot {
-    identity: Option<TurnCheckpointIdentity>,
-    preparation: TurnPreparationSnapshot,
-    request: TurnCheckpointRequest,
-    lane: Option<TurnLaneExecutionSnapshot>,
-    reply: Option<TurnReplyCheckpoint>,
-    finalization: TurnFinalizationCheckpoint,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-struct TurnCheckpointIdentity {
-    user_input_sha256: String,
-    assistant_reply_sha256: String,
-    user_input_chars: usize,
-    assistant_reply_chars: usize,
-}
-
-impl TurnCheckpointIdentity {
-    fn from_turn(user_input: &str, assistant_reply: &str) -> Self {
-        Self {
-            user_input_sha256: sha256_hex(user_input),
-            assistant_reply_sha256: sha256_hex(assistant_reply),
-            user_input_chars: user_input.chars().count(),
-            assistant_reply_chars: assistant_reply.chars().count(),
-        }
-    }
-
-    fn matches_turn(&self, user_input: &str, assistant_reply: &str) -> bool {
-        self.user_input_chars == user_input.chars().count()
-            && self.assistant_reply_chars == assistant_reply.chars().count()
-            && self.user_input_sha256 == sha256_hex(user_input)
-            && self.assistant_reply_sha256 == sha256_hex(assistant_reply)
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize)]
-struct TurnPreparationSnapshot {
-    lane: ExecutionLane,
-    max_tool_steps: usize,
-    raw_tool_output_requested: bool,
-    context_message_count: usize,
-    context_fingerprint_sha256: String,
-    estimated_tokens: Option<usize>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize)]
-#[serde(tag = "kind", rename_all = "snake_case")]
-enum TurnCheckpointRequest {
-    Continue { tool_intents: usize },
-    FinalizeInlineProviderError,
-    ReturnError,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize)]
-struct TurnLaneExecutionSnapshot {
-    lane: ExecutionLane,
-    had_tool_intents: bool,
-    raw_tool_output_requested: bool,
-    result_kind: TurnCheckpointResultKind,
-    safe_lane_terminal_route: Option<SafeLaneFailureRoute>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
-#[serde(rename_all = "snake_case")]
-enum TurnCheckpointResultKind {
-    FinalText,
-    Streaming,
-    NeedsApproval,
-    ToolDenied,
-    ToolError,
-    ProviderError,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize)]
-struct TurnReplyCheckpoint {
-    decision: ReplyResolutionMode,
-    followup_kind: Option<ToolDrivenFollowupKind>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
-#[serde(tag = "kind", rename_all = "snake_case")]
-enum TurnFinalizationCheckpoint {
-    PersistReply {
-        persistence_mode: ReplyPersistenceMode,
-        runs_after_turn: bool,
-        attempts_context_compaction: bool,
-    },
-    ReturnError,
-}
-
-impl TurnFinalizationCheckpoint {
-    fn persist_reply(persistence_mode: ReplyPersistenceMode) -> Self {
-        Self::PersistReply {
-            persistence_mode,
-            runs_after_turn: true,
-            attempts_context_compaction: true,
-        }
-    }
-
-    fn persistence_mode(self) -> Option<ReplyPersistenceMode> {
-        match self {
-            Self::PersistReply {
-                persistence_mode, ..
-            } => Some(persistence_mode),
-            Self::ReturnError => None,
-        }
-    }
-
-    fn runs_after_turn(self) -> bool {
-        match self {
-            Self::PersistReply {
-                runs_after_turn, ..
-            } => runs_after_turn,
-            Self::ReturnError => false,
-        }
-    }
-
-    fn attempts_context_compaction(self) -> bool {
-        match self {
-            Self::PersistReply {
-                attempts_context_compaction,
-                ..
-            } => attempts_context_compaction,
-            Self::ReturnError => false,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
-#[serde(rename_all = "snake_case")]
-enum TurnCheckpointStage {
-    PostPersist,
-    Finalized,
-    FinalizationFailed,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
-#[serde(rename_all = "snake_case")]
-enum TurnCheckpointProgressStatus {
-    Pending,
-    Skipped,
-    Completed,
-    Failed,
-    FailedOpen,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
-struct TurnCheckpointFinalizationProgress {
-    after_turn: TurnCheckpointProgressStatus,
-    compaction: TurnCheckpointProgressStatus,
-}
-
-impl TurnCheckpointFinalizationProgress {
-    fn pending(checkpoint: &TurnCheckpointSnapshot) -> Self {
-        Self {
-            after_turn: if checkpoint.finalization.runs_after_turn() {
-                TurnCheckpointProgressStatus::Pending
-            } else {
-                TurnCheckpointProgressStatus::Skipped
-            },
-            compaction: if checkpoint.finalization.attempts_context_compaction() {
-                TurnCheckpointProgressStatus::Pending
-            } else {
-                TurnCheckpointProgressStatus::Skipped
-            },
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum ContextCompactionOutcome {
-    Skipped,
-    Completed,
-    FailedOpen,
-}
-
-impl ContextCompactionOutcome {
-    fn checkpoint_status(self) -> TurnCheckpointProgressStatus {
-        match self {
-            Self::Skipped => TurnCheckpointProgressStatus::Skipped,
-            Self::Completed => TurnCheckpointProgressStatus::Completed,
-            Self::FailedOpen => TurnCheckpointProgressStatus::FailedOpen,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
-#[serde(rename_all = "snake_case")]
-enum TurnCheckpointFailureStep {
-    AfterTurn,
-    Compaction,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize)]
-struct TurnCheckpointFailure {
-    step: TurnCheckpointFailureStep,
-    error: String,
 }
 
 #[derive(Debug, Clone, PartialEq, Eq)]
@@ -1500,13 +1013,6 @@ impl SafeLaneTurnOutcome {
             result,
             terminal_route: Some(terminal_route),
         }
-    }
-}
-
-fn turn_reply_checkpoint(phase: &ToolDrivenReplyPhase) -> TurnReplyCheckpoint {
-    TurnReplyCheckpoint {
-        decision: phase.resolution_mode(),
-        followup_kind: phase.followup_kind(),
     }
 }
 
@@ -3393,41 +2899,6 @@ async fn resolve_provider_turn_reply<R: ConversationRuntime + ?Sized>(
     }
 }
 
-fn turn_checkpoint_result_kind(result: &TurnResult) -> TurnCheckpointResultKind {
-    match result {
-        TurnResult::FinalText(_) => TurnCheckpointResultKind::FinalText,
-        TurnResult::StreamingText(_) | TurnResult::StreamingDone(_) => {
-            TurnCheckpointResultKind::Streaming
-        }
-        TurnResult::NeedsApproval(_) => TurnCheckpointResultKind::NeedsApproval,
-        TurnResult::ToolDenied(_) => TurnCheckpointResultKind::ToolDenied,
-        TurnResult::ToolError(_) => TurnCheckpointResultKind::ToolError,
-        TurnResult::ProviderError(_) => TurnCheckpointResultKind::ProviderError,
-    }
-}
-
-fn format_turn_checkpoint_progress_status(status: TurnCheckpointProgressStatus) -> &'static str {
-    match status {
-        TurnCheckpointProgressStatus::Pending => "pending",
-        TurnCheckpointProgressStatus::Skipped => "skipped",
-        TurnCheckpointProgressStatus::Completed => "completed",
-        TurnCheckpointProgressStatus::Failed => "failed",
-        TurnCheckpointProgressStatus::FailedOpen => "failed_open",
-    }
-}
-
-fn format_analytics_turn_checkpoint_progress_status(
-    status: AnalyticsTurnCheckpointProgressStatus,
-) -> &'static str {
-    match status {
-        AnalyticsTurnCheckpointProgressStatus::Pending => "pending",
-        AnalyticsTurnCheckpointProgressStatus::Skipped => "skipped",
-        AnalyticsTurnCheckpointProgressStatus::Completed => "completed",
-        AnalyticsTurnCheckpointProgressStatus::Failed => "failed",
-        AnalyticsTurnCheckpointProgressStatus::FailedOpen => "failed_open",
-    }
-}
-
 #[cfg(test)]
 fn build_turn_reply_followup_messages(
     base_messages: &[Value],
@@ -3559,190 +3030,6 @@ async fn emit_discovery_first_event<R: ConversationRuntime + ?Sized>(
     }
 }
 
-async fn persist_turn_checkpoint_event<R: ConversationRuntime + ?Sized>(
-    runtime: &R,
-    session_id: &str,
-    checkpoint: &TurnCheckpointSnapshot,
-    stage: TurnCheckpointStage,
-    progress: TurnCheckpointFinalizationProgress,
-    failure: Option<TurnCheckpointFailure>,
-    binding: ConversationRuntimeBinding<'_>,
-) -> CliResult<()> {
-    let checkpoint = serde_json::to_value(checkpoint)
-        .map_err(|error| format!("serialize turn checkpoint failed: {error}"))?;
-    persist_turn_checkpoint_event_value(
-        runtime,
-        session_id,
-        &checkpoint,
-        stage,
-        progress,
-        failure,
-        binding,
-    )
-    .await
-}
-
-async fn persist_turn_checkpoint_event_value<R: ConversationRuntime + ?Sized>(
-    runtime: &R,
-    session_id: &str,
-    checkpoint: &Value,
-    stage: TurnCheckpointStage,
-    progress: TurnCheckpointFinalizationProgress,
-    failure: Option<TurnCheckpointFailure>,
-    binding: ConversationRuntimeBinding<'_>,
-) -> CliResult<()> {
-    persist_conversation_event(
-        runtime,
-        session_id,
-        "turn_checkpoint",
-        json!({
-            "schema_version": 1,
-            "stage": stage,
-            "checkpoint": checkpoint,
-            "finalization_progress": progress,
-            "failure": failure,
-        }),
-        binding,
-    )
-    .await
-}
-
-fn recover_latest_turn_pair(messages: &[Value]) -> Option<(String, String)> {
-    let assistant_index = messages.iter().rposition(|message| {
-        message.get("role").and_then(Value::as_str) == Some("assistant")
-            && message.get("content").and_then(Value::as_str).is_some()
-    })?;
-    let assistant_reply = messages
-        .get(assistant_index)?
-        .get("content")
-        .and_then(Value::as_str)?
-        .to_owned();
-    let user_input = messages
-        .get(..assistant_index)?
-        .iter()
-        .rposition(|message| {
-            message.get("role").and_then(Value::as_str) == Some("user")
-                && message.get("content").and_then(Value::as_str).is_some()
-        })
-        .and_then(|index| {
-            messages
-                .get(index)
-                .and_then(|message| message.get("content"))
-                .and_then(Value::as_str)
-        })
-        .map(ToOwned::to_owned)?;
-    Some((user_input, assistant_reply))
-}
-
-fn load_turn_checkpoint_identity(checkpoint: &Value) -> Option<TurnCheckpointIdentity> {
-    checkpoint
-        .get("identity")
-        .cloned()
-        .and_then(|identity| serde_json::from_value(identity).ok())
-}
-
-fn sha256_hex(input: &str) -> String {
-    format!("{:x}", Sha256::digest(input.as_bytes()))
-}
-
-fn checkpoint_context_fingerprint_sha256(messages: &[Value]) -> String {
-    let serialized = Value::Array(messages.to_vec()).to_string();
-    format!("{:x}", Sha256::digest(serialized.as_bytes()))
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize, Default)]
-struct TurnCheckpointRepairPreparation {
-    #[serde(default)]
-    context_message_count: Option<usize>,
-    #[serde(default)]
-    context_fingerprint_sha256: Option<String>,
-    #[serde(default)]
-    estimated_tokens: Option<usize>,
-}
-
-fn load_turn_checkpoint_repair_preparation(
-    checkpoint: &Value,
-) -> Result<Option<TurnCheckpointRepairPreparation>, TurnCheckpointTailRepairReason> {
-    let Some(preparation) = checkpoint.get("preparation") else {
-        return Ok(None);
-    };
-    serde_json::from_value(preparation.clone())
-        .map(Some)
-        .map_err(|_error| TurnCheckpointTailRepairReason::CheckpointPreparationMalformed)
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct TurnCheckpointRepairResumeInput {
-    user_input: String,
-    assistant_reply: String,
-    messages: Vec<Value>,
-    estimated_tokens: Option<usize>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-enum TurnCheckpointTailRuntimeEligibility {
-    NotNeeded {
-        action: TurnCheckpointRecoveryAction,
-        reason: TurnCheckpointTailRepairReason,
-    },
-    Manual {
-        action: TurnCheckpointRecoveryAction,
-        reason: TurnCheckpointTailRepairReason,
-        source: TurnCheckpointTailRepairSource,
-    },
-    Runnable {
-        action: TurnCheckpointRecoveryAction,
-        plan: TurnCheckpointRepairPlan,
-        resume_input: TurnCheckpointRepairResumeInput,
-    },
-}
-
-impl TurnCheckpointRepairResumeInput {
-    fn from_assembled_context(
-        assembled: AssembledConversationContext,
-        checkpoint: &Value,
-    ) -> Result<Self, TurnCheckpointTailRepairReason> {
-        let repair_preparation = load_turn_checkpoint_repair_preparation(checkpoint)?;
-        let messages = assembled.messages;
-        let Some((user_input, assistant_reply)) = recover_latest_turn_pair(&messages) else {
-            return Err(TurnCheckpointTailRepairReason::VisibleTurnPairMissing);
-        };
-        let Some((_, pre_assistant_messages)) = messages.split_last() else {
-            return Err(TurnCheckpointTailRepairReason::VisibleTurnPairMissing);
-        };
-        let Some(identity) = load_turn_checkpoint_identity(checkpoint) else {
-            return Err(TurnCheckpointTailRepairReason::CheckpointIdentityMissing);
-        };
-        if !identity.matches_turn(&user_input, &assistant_reply) {
-            return Err(TurnCheckpointTailRepairReason::CheckpointIdentityMismatch);
-        }
-        if let Some(expected_context_message_count) = repair_preparation
-            .as_ref()
-            .and_then(|preparation| preparation.context_message_count)
-            && pre_assistant_messages.len() != expected_context_message_count
-        {
-            return Err(TurnCheckpointTailRepairReason::CheckpointPreparationMismatch);
-        }
-        if let Some(expected_context_fingerprint_sha256) = repair_preparation
-            .as_ref()
-            .and_then(|preparation| preparation.context_fingerprint_sha256.as_deref())
-            && checkpoint_context_fingerprint_sha256(pre_assistant_messages)
-                != expected_context_fingerprint_sha256
-        {
-            return Err(TurnCheckpointTailRepairReason::CheckpointPreparationFingerprintMismatch);
-        }
-
-        Ok(Self {
-            user_input,
-            assistant_reply,
-            messages,
-            estimated_tokens: repair_preparation
-                .and_then(|preparation| preparation.estimated_tokens)
-                .or(assembled.estimated_tokens),
-        })
-    }
-}
-
 #[cfg(feature = "memory-sqlite")]
 async fn repair_turn_checkpoint_tail_entry<R: ConversationRuntime + ?Sized>(
     config: &LoongClawConfig,
@@ -3820,9 +3107,9 @@ async fn repair_turn_checkpoint_tail_entry<R: ConversationRuntime + ?Sized>(
         match runtime
             .after_turn(
                 session_id,
-                &resume_input.user_input,
-                &resume_input.assistant_reply,
-                &resume_input.messages,
+                resume_input.user_input(),
+                resume_input.assistant_reply(),
+                resume_input.messages(),
                 kernel_ctx,
             )
             .await
@@ -3861,8 +3148,8 @@ async fn repair_turn_checkpoint_tail_entry<R: ConversationRuntime + ?Sized>(
             config,
             runtime,
             session_id,
-            &resume_input.messages,
-            resume_input.estimated_tokens,
+            resume_input.messages(),
+            resume_input.estimated_tokens(),
             binding,
         )
         .await
@@ -4003,20 +3290,6 @@ async fn probe_turn_checkpoint_tail_runtime_gate_entry_with_limit<
     };
     probe_turn_checkpoint_tail_runtime_gate_entry(config, runtime, session_id, &entry, binding)
         .await
-}
-
-fn restore_analytics_turn_checkpoint_progress_status(
-    status: AnalyticsTurnCheckpointProgressStatus,
-) -> TurnCheckpointProgressStatus {
-    match status {
-        AnalyticsTurnCheckpointProgressStatus::Pending => TurnCheckpointProgressStatus::Pending,
-        AnalyticsTurnCheckpointProgressStatus::Skipped => TurnCheckpointProgressStatus::Skipped,
-        AnalyticsTurnCheckpointProgressStatus::Completed => TurnCheckpointProgressStatus::Completed,
-        AnalyticsTurnCheckpointProgressStatus::Failed => TurnCheckpointProgressStatus::Failed,
-        AnalyticsTurnCheckpointProgressStatus::FailedOpen => {
-            TurnCheckpointProgressStatus::FailedOpen
-        }
-    }
 }
 
 async fn finalize_provider_turn_reply<R: ConversationRuntime + ?Sized>(
@@ -6539,7 +5812,7 @@ fn safe_lane_backpressure_budget(config: &LoongClawConfig) -> Option<SafeLaneBac
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
-struct SafeLaneFailureRoute {
+pub(super) struct SafeLaneFailureRoute {
     decision: SafeLaneFailureRouteDecision,
     reason: SafeLaneFailureRouteReason,
     source: SafeLaneFailureRouteSource,
